% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campaigns}
\alias{campaigns}
\title{House/Senate Current Campaigns}
\format{
A data frame with 1898 rows and 25 columns, providing financial
information for each House/Senate campaign:
\describe{
\item{cand_id}{A code assigned to a candidate by the FEC}
\item{cand_name}{Candidate name}
\item{cand_ici}{Incumbent challenger status with designation:
C = Challenger, I = Incumbent, O = Open Seat}
\item{pty_cd}{Party code}
\item{cand_pty_affiliation}{The political party affiliation reported by the candidate}
\item{ttl_receipts}{Total receipts}
\item{trans_from_auth}{Transfers from authorized committees}
\item{ttl_disb}{Total disbursements}
\item{trans_to_auth}{Transfers to authorized committees}
\item{coh_bop}{Beginning cash}
\item{coh_cop}{Ending cash}
\item{cand_contrib}{Contributions from candidate}
\item{cand_loans}{Loans from candidate}
\item{other_loans}{Other loans}
\item{cand_loan_repay}{Candidate loan repayments}
\item{other_loan_repay}{Other loan repayments}
\item{debts_owed_by}{Debts owed by}
\item{ttl_indiv_contrib}{Total individual contribution}
\item{cand_office_st}{Candidate's state}
\item{cand_office_district}{Candidate's district}
\item{other_pol_cmte_contrib}{Contributions from other political committees}
\item{pol_pty_contrib}{Contributions from party committees}
\item{cvg_end_dt}{Coverage end date (YYYY-MM-DD)}
\item{indiv_refunds}{Refunds to individuals}
\item{cmte_refunds}{Refunds to committees}
}
}
\source{
\url{https://transition.fec.gov/general/FederalElections2016.shtml}, updated 2020-03-02
}
\usage{
campaigns
}
\description{
Summary financial information for each campaign.
}
\keyword{datasets}
