% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pac}
\alias{pac}
\title{Political Action Committee (PAC) and Party Summary Financial Information}
\format{
A data frame with 12049 rows and 27 columns:
\describe{
\item{cmte_id}{A 9-character alpha-numeric code assigned to a committee by the Federal Election Commission}
\item{cmte_nm}{Name of the committee}
\item{cmte_tp}{Committee Type, listed here: \url{https://www.fec.gov/campaign-finance-data/committee-type-code-descriptions/}}
\item{cmte_dsgn}{A = Authorized by a candidate

B = Lobbyist/Registrant PAC

D = Leadership PAC

J = Joint fundraiser

P = Principal campaign committee of a candidate

U = Unauthorized}

\item{cmte_filing_freq}{Filing frequency: A = Administratively terminated,

D = Debt

M = Monthly filer

Q = Quarterly filer

T = Terminated

W = Waived}

\item{ttl_receipts}{Total receipts}
\item{trans_from_aff}{Transfers from affiliates}
\item{indv_contrib}{Contributions from individuals}
\item{other_pol_cmte_contrib}{Contributions from other political committees}
\item{cand_contrib}{Contributions from candidate}
\item{cand_loans}{Candidate loans}
\item{ttl_loans_received}{Total loans received}
\item{ttl_disb}{Total disbursements}
\item{tranf_to_aff}{Transfers to affiliates}
\item{indv_refunds}{Refunds to individuals}
\item{other_pol_cmte_refunds}{Refunds to other political committees}
\item{cand_loan_repay}{Candidate loan repayments}
\item{loan_repay}{Loan repayments}
\item{coh_bop}{Cash beginning of period}
\item{coh_cop}{Cash close of period}
\item{debts_owed_by}{Debts owed by}
\item{nonfed_trans_received}{Nonfederal transfers received}
\item{contrib_to_other_cmte}{Contributions to other committees}
\item{ind_exp}{Independent expenditures}
\item{pty_coord_exp}{Party coordinated expenditures}
\item{nonfed_share_exp}{Nonfederal share expenditures}
\item{cvg_end_dt}{Coverage end date (YYYY-MM-DD)}
}
}
\source{
\url{https://www.fec.gov/campaign-finance-data/pac-and-party-summary-file-description/}, updated 2020-03-02
}
\usage{
pac
}
\description{
PAC and party summary finanicial information
}
\keyword{datasets}
