% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_expenditures.R
\name{read_all_expenditures}
\alias{read_all_expenditures}
\title{Operating Expenditures}
\usage{
read_all_expenditures(n_max = Inf)
}
\arguments{
\item{n_max}{integer specifying the max amount of entries in the dataset, defaults to the possible maximum}
}
\value{
The entire dataframe. More information about variables is at \code{?expenditures}.
}
\description{
\code{read_all_expenditures} returns a dataframe about the committees master data
}
\examples{
\dontrun{read_all_expenditures()}
\dontrun{read_all_expenditures(n_max = 250)}
}
