% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transactions.R
\name{read_all_transactions}
\alias{read_all_transactions}
\title{Any Transaction From One Committee To Another}
\usage{
read_all_transactions(n_max = Inf)
}
\arguments{
\item{n_max}{integer specifying the max amount of entries in the dataset, defaults to the possible maximum}
}
\value{
The entire dataframe. More information about variables is at \code{?transactions}.
}
\description{
\code{read_all_transactions} returns a dataframe about the committees master data
}
\examples{
\dontrun{read_all_transactions()}
\dontrun{read_all_transactions(n_max = 250)}
}
