% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results_house}
\alias{results_house}
\alias{results_senate}
\alias{results_president}
\title{General Federal Election Results}
\format{
3 data frames with information about the House, Senate, and Presidential Election results:
\describe{
\item{state}{The state of the votes}
\item{district_id}{District ID}
\item{cand_id}{A code assigned to a candidate by the FEC}
\item{incumbent}{A logical variable: TRUE if incumbent, FALSE if not}
\item{party}{Party of the candidate}
\item{primary_votes}{Number of votes in the primary electiom}
\item{primary_percent}{Percentage of votes in the primary election}
\item{runoff_votes}{Number of runoff votes}
\item{runoff_percent}{Percentage of runoff votes}
\item{general_votes}{Number of votes in the general election}
\item{general_percent}{Percentage of votes in the general election}
\item{won}{A logical variable: TRUE if candidate won, FALSE if not}
\item{footnotes}{Footnotes}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 377 rows and 10 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 619 rows and 6 columns.
}
\source{
\url{https://transition.fec.gov/general/FederalElections2016.shtml}, updated 2020-03-02
}
\usage{
results_house

results_senate

results_president
}
\description{
2016 House General Election Results

2016 Senate General Election Results

2016 Presidential General Election Results
}
\keyword{datasets}
